get image

+++ main.py
@@ -27,7 +27,7 @@ if __name__ == '__main__':
     image = Image.open('smallnorb_export/train/012149_car_04_rt.jpg')
     # summarize some details about the image
     image.load()
-    data = np.asarray(image, dtype="int32")
+    data = np.asarray(image, dtype="float32")
     print(image.format)
     print(image.mode)
     print(image.size)
